Made use of card as a separate input field
how do we customize the input?

the input works by pushing object into the array, we just need to create an object class by being consistent in the naming of subfields, i.e. in this case, dessert have a name, price and desc. That way, the same structure would be passed into the data structure